+++ tests/text/test_rouge.py
@@ -12,11 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from functools import partial
 from typing import List
 
 import pytest
-import torch
 
+from tests.text.helpers import INPUT_ORDER, TextTester
 from torchmetrics.functional.text.rouge import rouge_score
 from torchmetrics.text.rouge import ROUGEScore
 from torchmetrics.utilities.imports import _NLTK_AVAILABLE, _ROUGE_SCORE_AVAILABLE
@@ -32,27 +33,34 @@ ROUGE_KEYS = ("rouge1", "rouge2", "rougeL", "rougeLsum")
 SINGLE_SENTENCE_EXAMPLE_PREDS = "The quick brown fox jumps over the lazy dog"
 SINGLE_SENTENCE_EXAMPLE_TARGET = "The quick brown dog jumps on the log."
 
-PREDS = "My name is John".split()
-TARGETS = "Is your name John".split()
+PREDS = "My name is John"
+TARGETS = "Is your name John"
 
 
-BATCHES_RS_PREDS = [SINGLE_SENTENCE_EXAMPLE_PREDS]
-BATCHES_RS_PREDS.extend(PREDS)
-BATCHES_RS_TARGETS = [SINGLE_SENTENCE_EXAMPLE_TARGET]
-BATCHES_RS_TARGETS.extend(TARGETS)
-
-BATCHES = [
-    dict(preds=[SINGLE_SENTENCE_EXAMPLE_PREDS], targets=[SINGLE_SENTENCE_EXAMPLE_TARGET]),
-    dict(preds=PREDS, targets=TARGETS),
-]
+BATCHES_1 = {
+    "preds": [["the cat was under the bed"], ["the cat was found under the bed"]],
+    "targets": [["the cat was found under the bed"], ["the tiny little cat was found under the big funny bed "]],
+}
 
 
-def _compute_rouge_score(preds: List[str], targets: List[str], use_stemmer: bool):
+BATCHES_2 = {
+    "preds": [["The quick brown fox jumps over the lazy dog"], ["My name is John"]],
+    "targets": [["The quick brown dog jumps on the log."], ["Is your name John"]],
+}
+
+
+def _compute_rouge_score(preds: List[str], targets: List[str], use_stemmer: bool, rouge_level: str, metric: str):
+    if isinstance(preds, str):
+        preds = [preds]
+    if isinstance(targets, str):
+        targets = [targets]
     scorer = RougeScorer(ROUGE_KEYS, use_stemmer=use_stemmer)
     aggregator = BootstrapAggregator()
     for pred, target in zip(preds, targets):
         aggregator.add_scores(scorer.score(target, pred))
-    return aggregator.aggregate()
+    rs_scores = aggregator.aggregate()
+    rs_result = getattr(rs_scores[rouge_level].mid, metric)
+    return rs_result
 
 
 @pytest.mark.skipif(not _NLTK_AVAILABLE, reason="test requires nltk")
@@ -73,108 +81,49 @@ def _compute_rouge_score(preds: List[str], targets: List[str], use_stemmer: bool
         pytest.param("rougeLsum_fmeasure", False),
     ],
 )
-def test_rouge_metric_functional_single_sentence(pl_rouge_metric_key, use_stemmer):
-    rouge_level, metric = pl_rouge_metric_key.split("_")
-
-    scorer = RougeScorer(ROUGE_KEYS, use_stemmer=use_stemmer)
-    rs_scores = scorer.score(SINGLE_SENTENCE_EXAMPLE_TARGET, SINGLE_SENTENCE_EXAMPLE_PREDS)
-    rs_result = torch.tensor(getattr(rs_scores[rouge_level], metric), dtype=torch.float32)
-
-    pl_output = rouge_score([SINGLE_SENTENCE_EXAMPLE_PREDS], [SINGLE_SENTENCE_EXAMPLE_TARGET], use_stemmer=use_stemmer)
-
-    assert torch.allclose(pl_output[pl_rouge_metric_key], rs_result)
-
-
-@pytest.mark.skipif(not _NLTK_AVAILABLE, reason="test requires nltk")
 @pytest.mark.parametrize(
-    ["pl_rouge_metric_key", "use_stemmer"],
+    ["preds", "targets"],
     [
-        pytest.param("rouge1_precision", True),
-        pytest.param("rouge1_recall", True),
-        pytest.param("rouge1_fmeasure", False),
-        pytest.param("rouge2_precision", False),
-        pytest.param("rouge2_recall", True),
-        pytest.param("rouge2_fmeasure", True),
-        pytest.param("rougeL_precision", False),
-        pytest.param("rougeL_recall", False),
-        pytest.param("rougeL_fmeasure", True),
-        pytest.param("rougeLsum_precision", True),
-        pytest.param("rougeLsum_recall", False),
-        pytest.param("rougeLsum_fmeasure", False),
+        pytest.param(BATCHES_1["preds"], BATCHES_1["targets"]),
+        pytest.param(BATCHES_2["preds"], BATCHES_2["targets"]),
     ],
 )
-def test_rouge_metric_functional(pl_rouge_metric_key, use_stemmer):
-    rouge_level, metric = pl_rouge_metric_key.split("_")
+class TestROUGEScore(TextTester):
+    @pytest.mark.parametrize("ddp", [False, True])
+    @pytest.mark.parametrize("dist_sync_on_step", [False, True])
+    def test_rouge_score_class(self, ddp, dist_sync_on_step, preds, targets, pl_rouge_metric_key, use_stemmer):
+        metric_args = {"use_stemmer": use_stemmer}
 
-    rs_scores = _compute_rouge_score(PREDS, TARGETS, use_stemmer=use_stemmer)
-    rs_result = torch.tensor(getattr(rs_scores[rouge_level].mid, metric), dtype=torch.float32)
+        rouge_level, metric = pl_rouge_metric_key.split("_")
+        rouge_metric = partial(_compute_rouge_score, use_stemmer=use_stemmer, rouge_level=rouge_level, metric=metric)
 
-    pl_output = rouge_score(PREDS, TARGETS, use_stemmer=use_stemmer)
+        self.run_class_metric_test(
+            ddp=ddp,
+            preds=preds,
+            targets=targets,
+            metric_class=ROUGEScore,
+            sk_metric=rouge_metric,
+            dist_sync_on_step=dist_sync_on_step,
+            metric_args=metric_args,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+            key=pl_rouge_metric_key,
+        )
 
-    assert torch.allclose(pl_output[pl_rouge_metric_key], rs_result)
+    def test_rouge_score_functional(self, preds, targets, pl_rouge_metric_key, use_stemmer):
+        metric_args = {"use_stemmer": use_stemmer}
 
+        rouge_level, metric = pl_rouge_metric_key.split("_")
+        rouge_metric = partial(_compute_rouge_score, use_stemmer=use_stemmer, rouge_level=rouge_level, metric=metric)
 
-@pytest.mark.skipif(not _NLTK_AVAILABLE, reason="test requires nltk")
-@pytest.mark.parametrize(
-    ["pl_rouge_metric_key", "use_stemmer"],
-    [
-        pytest.param("rouge1_precision", True),
-        pytest.param("rouge1_recall", True),
-        pytest.param("rouge1_fmeasure", False),
-        pytest.param("rouge2_precision", False),
-        pytest.param("rouge2_recall", True),
-        pytest.param("rouge2_fmeasure", True),
-        pytest.param("rougeL_precision", False),
-        pytest.param("rougeL_recall", False),
-        pytest.param("rougeL_fmeasure", True),
-        pytest.param("rougeLsum_precision", True),
-        pytest.param("rougeLsum_recall", False),
-        pytest.param("rougeLsum_fmeasure", False),
-    ],
-)
-def test_rouge_metric_class(pl_rouge_metric_key, use_stemmer):
-    rouge_level, metric = pl_rouge_metric_key.split("_")
-
-    scorer = RougeScorer(ROUGE_KEYS, use_stemmer=use_stemmer)
-    rs_scores = scorer.score(SINGLE_SENTENCE_EXAMPLE_TARGET, SINGLE_SENTENCE_EXAMPLE_PREDS)
-    rs_result = torch.tensor(getattr(rs_scores[rouge_level], metric), dtype=torch.float32)
-
-    rouge = ROUGEScore(use_stemmer=use_stemmer)
-    pl_output = rouge([SINGLE_SENTENCE_EXAMPLE_PREDS], [SINGLE_SENTENCE_EXAMPLE_TARGET])
-
-    assert torch.allclose(pl_output[pl_rouge_metric_key], rs_result)
-
-
-@pytest.mark.skipif(not _NLTK_AVAILABLE, reason="test requires nltk")
-@pytest.mark.parametrize(
-    ["pl_rouge_metric_key", "use_stemmer"],
-    [
-        pytest.param("rouge1_precision", True),
-        pytest.param("rouge1_recall", True),
-        pytest.param("rouge1_fmeasure", False),
-        pytest.param("rouge2_precision", False),
-        pytest.param("rouge2_recall", True),
-        pytest.param("rouge2_fmeasure", True),
-        pytest.param("rougeL_precision", False),
-        pytest.param("rougeL_recall", False),
-        pytest.param("rougeL_fmeasure", True),
-        pytest.param("rougeLsum_precision", True),
-        pytest.param("rougeLsum_recall", False),
-        pytest.param("rougeLsum_fmeasure", False),
-    ],
-)
-def test_rouge_metric_class_batches(pl_rouge_metric_key, use_stemmer):
-    rouge_level, metric = pl_rouge_metric_key.split("_")
-
-    rs_scores = _compute_rouge_score(BATCHES_RS_PREDS, BATCHES_RS_TARGETS, use_stemmer=use_stemmer)
-    rs_result = torch.tensor(getattr(rs_scores[rouge_level].mid, metric), dtype=torch.float32)
-
-    rouge = ROUGEScore(use_stemmer=use_stemmer)
-    for batch in BATCHES:
-        rouge.update(batch["preds"], batch["targets"])
-    pl_output = rouge.compute()
-
-    assert torch.allclose(pl_output[pl_rouge_metric_key], rs_result)
+        self.run_functional_metric_test(
+            preds,
+            targets,
+            metric_functional=rouge_score,
+            sk_metric=rouge_metric,
+            metric_args=metric_args,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+            key=pl_rouge_metric_key,
+        )
 
 
 def test_rouge_metric_raises_errors_and_warnings():