+++ tests/text/test_wer.py
@@ -1,68 +1,75 @@
+from typing import Callable, List, Union
+
 import pytest
 
+from tests.text.helpers import INPUT_ORDER, TextTester
 from torchmetrics.utilities.imports import _JIWER_AVAILABLE
 
 if _JIWER_AVAILABLE:
     from jiwer import compute_measures
+else:
+    compute_measures = Callable
 
 from torchmetrics.functional.text.wer import wer
 from torchmetrics.text.wer import WER
 
-PREDICTION1 = "hello world"
-REFERENCE1 = "hello world"
+BATCHES_1 = {"preds": [["hello world"], ["what a day"]], "targets": [["hello world"], ["what a wonderful day"]]}
 
-PREDICTION2 = "what a day"
-REFERENCE2 = "what a wonderful day"
+BATCHES_2 = {
+    "preds": [
+        ["i like python", "what you mean or swallow"],
+        ["hello duck", "i like python"],
+    ],
+    "targets": [
+        ["i like monthy python", "what do you mean, african or european swallow"],
+        ["hello world", "i like monthy python"],
+    ],
+}
 
-BATCH_PREDICTIONS = [PREDICTION1, PREDICTION2]
-BATCH_REFERENCES = [REFERENCE1, REFERENCE2]
+
+def _compute_wer_metric_jiwer(prediction: Union[str, List[str]], reference: Union[str, List[str]]):
+    return compute_measures(reference, prediction)["wer"]
 
 
+@pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
 @pytest.mark.parametrize(
-    "prediction,reference",
-    [(PREDICTION1, REFERENCE1), (PREDICTION2, REFERENCE2)],
+    ["preds", "targets"],
+    [
+        pytest.param(BATCHES_1["preds"], BATCHES_1["targets"]),
+        pytest.param(BATCHES_2["preds"], BATCHES_2["targets"]),
+    ],
 )
-def test_wer_functional_single_sentence(prediction, reference):
-    """Test functional with strings as inputs."""
-    pl_output = wer(prediction, reference)
-    jiwer_output = compute_measures(reference, prediction)["wer"]
-    assert pl_output == jiwer_output
+class TestWER(TextTester):
+    @pytest.mark.parametrize("ddp", [False, True])
+    @pytest.mark.parametrize("dist_sync_on_step", [False, True])
+    def test_wer_class(self, ddp, dist_sync_on_step, preds, targets):
 
+        self.run_class_metric_test(
+            ddp=ddp,
+            preds=preds,
+            targets=targets,
+            metric_class=WER,
+            sk_metric=_compute_wer_metric_jiwer,
+            dist_sync_on_step=dist_sync_on_step,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+        )
 
-def test_wer_functional_batch():
-    """Test functional with a batch of sentences."""
-    pl_output = wer(BATCH_PREDICTIONS, BATCH_REFERENCES)
-    jiwer_output = compute_measures(BATCH_REFERENCES, BATCH_PREDICTIONS)["wer"]
-    assert pl_output == jiwer_output
+    def test_wer_functional(self, preds, targets):
 
+        self.run_functional_metric_test(
+            preds,
+            targets,
+            metric_functional=wer,
+            sk_metric=_compute_wer_metric_jiwer,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+        )
 
-@pytest.mark.parametrize(
-    "prediction,reference",
-    [(PREDICTION1, REFERENCE1), (PREDICTION2, REFERENCE2)],
-)
-def test_wer_class_single_sentence(prediction, reference):
-    """Test class with strings as inputs."""
-    metric = WER()
-    metric.update(prediction, reference)
-    pl_output = metric.compute()
-    jiwer_output = compute_measures(reference, prediction)["wer"]
-    assert pl_output == jiwer_output
+    def test_wer_differentiability(self, preds, targets):
 
-
-def test_wer_class_batch():
-    """Test class with a batch of sentences."""
-    metric = WER()
-    metric.update(BATCH_PREDICTIONS, BATCH_REFERENCES)
-    pl_output = metric.compute()
-    jiwer_output = compute_measures(BATCH_REFERENCES, BATCH_PREDICTIONS)["wer"]
-    assert pl_output == jiwer_output
-
-
-def test_wer_class_batches():
-    """Test class with two batches of sentences."""
-    metric = WER()
-    for prediction, reference in zip(BATCH_PREDICTIONS, BATCH_REFERENCES):
-        metric.update(prediction, reference)
-    pl_output = metric.compute()
-    jiwer_output = compute_measures(BATCH_REFERENCES, BATCH_PREDICTIONS)["wer"]
-    assert pl_output == jiwer_output
+        self.run_differentiability_test(
+            preds=preds,
+            targets=targets,
+            metric_module=WER,
+            metric_functional=wer,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+        )