+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `MAP` (mean average precision) metric to image package ([#467](https://github.com/PyTorchLightning/metrics/pull/467))
 
 
+- Added `TextTester` class for robustly testing text metrics ([#450](https://github.com/PyTorchLightning/metrics/pull/450))
+
+
 ### Changed